refactor : topic domain data 레이어 리팩토링
