+++ lib/features/topic/data_source/local/topic_local_data_source_impl.dart
@@ -1,13 +1,11 @@
 import 'dart:convert';
 
-import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:collection/collection.dart';
 import 'package:flutter/services.dart';
 import 'package:hive/hive.dart';
 import 'package:techtalk/core/core.dart';
 import 'package:techtalk/features/topic/data_source/local/boxes/qna_box.dart';
 import 'package:techtalk/features/topic/data_source/local/boxes/qna_list_box.dart';
-import 'package:techtalk/features/topic/data_source/remote/models/topics_ref.dart';
 import 'package:techtalk/features/topic/topic.dart';
 
 class TopicLocalDataSourceImpl implements TopicLocalDataSource {
@@ -27,33 +25,6 @@ class TopicLocalDataSourceImpl implements TopicLocalDataSource {
         .cast<Map<String, dynamic>>()
         .map(TopicCategoryModel.fromJson)
         .toList();
-  }
-
-  @override
-  Future<List<TopicQnaModel>?> getQnas(
-    String topicId,
-  ) async {
-    final snapshot = await FirestoreTopicQuestionsRef.collection(topicId)
-        .get(const GetOptions(source: Source.cache));
-
-    if (snapshot.docs.isEmpty) {
-      return null;
-    }
-
-    return [
-      ...snapshot.docs.map((e) => e.data()),
-    ];
-  }
-
-  @override
-  Future<TopicQnaModel?> getQna(
-    String topicId,
-    String questionId,
-  ) async {
-    final snapshot = await FirestoreTopicQuestionsRef.doc(topicId, questionId)
-        .get(const GetOptions(source: Source.cache));
-
-    return snapshot.data();
   }
 
   @override
