+++ lib/features/topic/repositories/topic_repository.dart
@@ -4,14 +4,16 @@ import 'package:techtalk/features/topic/repositories/entities/wrong_answer_entit
 import 'package:techtalk/features/topic/topic.dart';
 
 abstract interface class TopicRepository {
-  Future<void> initStaticData();
-
-  Result<List<TopicCategoryEntity>> getTopicCategories();
-
+  ///
+  /// 문답 리스트 호출
+  ///
   Future<Result<List<QnaEntity>>> getTopicQnas(
     String topicId,
   );
 
+  ///
+  /// 단일 문답 호출
+  ///
   Future<Result<QnaEntity>> getTopicQna(
     String topicId,
     String questionId,
